// src/amazon_music/types/ShowLibraryHome.ts
export interface ShowLibraryHome {
    methods: (TemplateListBindMethod|InterfaceInterfaceMethod)[]
}

interface TemplateListBindMethod {
    interface: "TemplateListInterface.v1_0.BindTemplateMethod"
    template: Template
    queue: Queue64
    forced: boolean
}

interface InterfaceInterfaceMethod {
    interface: "InteractionInterface.v1_0.InvokeHttpSkillMethod"
    url: string
    clientInformation: any[]
    before: any[]
    after: any[]
    onSuccess: any[]
    onError: any[]
    queue: Queue
    forced: boolean
}

interface Template {
    interface: string
    emptyTemplateTitle: string
    emptyTemplateText: string
    emptyTemplateButtonText: string
    emptyTemplateButtonPrimaryLink: EmptyTemplateButtonPrimaryLink
    shouldHideSubNav: boolean
    widgets: Widget[]
    templateData: TemplateData6
    onEndOfWidgetsReached: any[]
    onCreated: OnCreated6[]
    onBound: any[]
    onViewed: OnViewed7[]
    onInteraction: any[]
    onLeave: any[]
    launchMode: string
    cacheSeconds: number
    isExploreButtonAndEmptyPromptNeeded: boolean
}
// ...
interface Queue {
    interface: string
    id: string
}
// ...
interface Queue64 {
    interface: string
    id: string
}