// src/amazon_music/types/SearchResult.ts
export interface SearchResult {
    methods: Method[]
}

interface Method {
    interface: string
    template: Template
    queue: Queue62
    forced: boolean
}

interface Template {
    interface: string
    widgets: Widget[]
    templateData: TemplateData5
    searchHeader: SearchHeader
    refinementHeader: RefinementHeader
    onEndOfWidgetsReached: any[]
    onCreated: OnCreated6[]
    onBound: any[]
    onViewed: OnViewed7[]
    onInteraction: any[]
    onLeave: any[]
    launchMode: string
    cacheSeconds: number
}
// ...
interface Queue62 {
    interface: string
    id: string
}